% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/terminator-package.R
\docType{package}
\name{terminator}
\alias{terminator}
\alias{terminator-package}
\title{Compute a single termiantor band}
\usage{
terminator(time, from = -180, to = 180, by = 0.1)
}
\arguments{
\item{time}{time (numeric from \code{POSIXct}) for the computation (bands are time-dependent)}

\item{from, to, by}{latitude sequence setup}
}
\value{
data frame
}
\description{
Returns a dataframe of latitude and longitude for the line that separates illuminated
day and dark night for any given time
}
\references{
\url{https://github.com/joergdietrich/Leaflet.Terminator/blob/master/L.Terminator.js},
\url{https://github.com/JoGall/terminator/blob/master/terminator.R}
}
\author{
Bob Rudis (bob@rud.is)
}
