% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminator_lat_lon.R
\name{terminator_lat_lon}
\alias{terminator_lat_lon}
\title{Generate a full set of terminator frames}
\usage{
terminator_lat_lon(day = Sys.Date(), hrs_from = 0, hrs_to = 23,
  hrs_by = 1, lon_from = -180, lon_to = 190, lon_by = 0.5)
}
\arguments{
\item{day, hrs_from, hrs_to, hrs_by}{setup for the hours sequence}

\item{lon_from, lon_to, lon_by}{latitude sequence setup}
}
\description{
Generate a full set of terminator frames
}
