% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminator_lat_lon.R
\name{terminator_lat_lon}
\alias{terminator_lat_lon}
\title{Generate a full set of terminator frames}
\usage{
terminator_lat_lon(day = Sys.Date(), from = 0, to = 23, by = 1)
}
\arguments{
\item{day, from, to, by}{setup for the hours sequence}
}
\description{
Generate a full set of terminator frames
}
